Move Rewriter and replace related methods to fast/rewriter
