+++ lib/fast.rb
@@ -2,6 +2,7 @@
 
 require 'fileutils'
 require 'astrolabe/builder'
+require_relative 'fast/rewriter'
 
 # suppress output to avoid parser gem warnings'
 def suppress_output
@@ -94,30 +95,6 @@ module Fast
       Matcher.new(pattern, ast, *args).match?
     end
 
-    # Replaces content based on a pattern.
-    # @param [Astrolabe::Node] ast with the current AST to search.
-    # @param [String] pattern with the expression to be targeting nodes.
-    # @param [Proc] replacement gives the [Rewriter] context in the block.
-    # @example
-    #   Fast.replace?(Fast.ast("a = 1"),"lvasgn") do |node|
-    #     replace(node.location.name, 'variable_renamed')
-    #   end # => variable_renamed = 1
-    # @return [String] with the new source code after apply the replacement
-    # @see Fast::Rewriter
-    def replace(pattern, ast, source = nil, &replacement)
-      buffer = Parser::Source::Buffer.new('replacement')
-      buffer.source = source || ast.loc.expression.source
-      to_replace = search(pattern, ast)
-      types = to_replace.grep(Parser::AST::Node).map(&:type).uniq
-
-      rewriter = Rewriter.new
-      rewriter.buffer = buffer
-      rewriter.search = pattern
-      rewriter.replacement = replacement
-      rewriter.replace_on(*types)
-      rewriter.rewrite(buffer, ast)
-    end
-
     # Search with pattern directly on file
     # @return [Array<Astrolabe::Node>] that matches the pattern
     def search_file(pattern, file)
@@ -148,21 +125,6 @@ module Fast
         .map { |f| { f => search_pattern.call(f) } }
         .reject { |results| results.values.all?(&:empty?) }
         .inject(&:merge!)
-    end
-
-    # Replaces the source of an {Fast#ast_from_file} with
-    # and the same source if the pattern does not match.
-    def replace_file(pattern, file, &replacement)
-      ast = ast_from_file(file)
-      replace(pattern, ast, IO.read(file), &replacement)
-    end
-
-    # Combines #replace_file output overriding the file if the output is different
-    # from the original file content.
-    def rewrite_file(pattern, file, &replacement)
-      previous_content = IO.read(file)
-      content = replace_file(pattern, file, &replacement)
-      File.open(file, 'w+') { |f| f.puts content } if content != previous_content
     end
 
     # Capture elements from searches in files. Keep in mind you need to use `$`
@@ -265,58 +227,6 @@ module Fast
         'nil'
       when Symbol, String, Numeric
         '_'
-      end
-    end
-  end
-
-  # Rewriter encapsulates {Rewriter#match_index} to allow
-  # {ExperimentFile.partial_replace} in a {Fast::ExperimentFile}.
-  # @see https://www.rubydoc.info/github/whitequark/parser/Parser/TreeRewriter
-  # @note the standalone class needs to combines {Rewriter#replace_on} to properly generate the `on_<node-type>` methods depending on the expression being used.
-  # @example Simple Rewriter
-  #    ast = Fast.ast("a = 1")
-  #    buffer = Parser::Source::Buffer.new('replacement')
-  #    buffer.source = ast.loc.expression.source
-  #    rewriter = Rewriter.new buffer
-  #    rewriter.search ='(lvasgn _ ...)'
-  #    rewriter.replacement =  -> (node) { replace(node.location.name, 'variable_renamed') }
-  #    rewriter.replace_on(:lvasgn)
-  #    rewriter.rewrite(buffer, ast) # => "variable_renamed = 1"
-  class Rewriter < Parser::TreeRewriter
-    # @return [Integer] with occurrence index
-    attr_reader :match_index
-    attr_accessor :buffer, :search, :replacement
-    def initialize(*args)
-      super
-      @match_index = 0
-    end
-
-    def match?(node)
-      Fast.match?(search, node)
-    end
-
-    # Generate methods for all affected types.
-    # @see Fast.replace
-    def replace_on(*types)
-      types.map do |type|
-        self.class.send :define_method, "on_#{type}" do |node|
-          if captures = match?(node) # rubocop:disable Lint/AssignmentInCondition
-            @match_index += 1
-            execute_replacement(node, captures)
-          end
-          super(node)
-        end
-      end
-    end
-
-    # Execute {#replacement} block
-    # @param [Astrolabe::Node] node that will be yield in the replacement block
-    # @param [Array<Object>, nil] captures are yield if {#replacement} take second argument.
-    def execute_replacement(node, captures)
-      if replacement.parameters.length == 1
-        instance_exec node, &replacement
-      else
-        instance_exec node, captures, &replacement
       end
     end
   end
